Extract most of EventRule code to EventCallHandler
